+++ src/Directory/Directory/Pages/js/pages/Account/Confirm/index.js
@@ -1,0 +1,2 @@
+export { default as ConfirmAccount } from "./ConfirmAccount";
+export { default as ResendConfirm } from "./ResendConfirm";
