+++ src/Directory/Directory/Pages/js/pages/Account/Login/components/LoginForm.js
@@ -2,8 +2,8 @@
 import { Formik, Form, Field } from "formik";
 import { postObjectAsFormData, constants } from "js-forms";
 import { Button, Flex, Stack, Box } from "@chakra-ui/core";
-import valSchema from "./login-form-validation";
-import CommonFormikInput from "./CommonFormikInput";
+import valSchema from "../login-form-validation";
+import CommonFormikInput from "../../../../components/CommonFormikInput";
 
 const LoginForm = ({ Username }) => {
   const aspForm = document.getElementById("asp-form");
