+++ src/Directory/Directory/Pages/js/pages/Account/Login/index.js
@@ -1,0 +1,2 @@
+export { default as Login } from "./Login";
+export { default as LoginRedirect } from "./LoginRedirect";
