+++ src/Directory/Directory/Pages/js/pages/Account/Logout/index.js
@@ -1,0 +1,2 @@
+export { default as ConfirmLogout } from "./ConfirmLogout";
+export { default as LogoutRedirect } from "./LogoutRedirect";
