+++ src/Directory/Directory/Pages/js/pages/Account/Register/RegisterResult.js
@@ -1,6 +1,6 @@
 import React from "react";
-import Layout from "../../Shared/Layout";
-import ConfirmationSent from "../components/ConfirmationSent";
+import Layout from "../../../components/Layout";
+import ConfirmationSent from "../../../components/ConfirmationSent";
 
 const RegisterResult = ({ Email }) => {
   return (
