+++ src/Directory/Directory/Pages/js/pages/Account/Register/components/RegisterForm.js
@@ -2,10 +2,10 @@ import React, { useState } from "react";
 import { Formik, Form, Field } from "formik";
 import { postObjectAsFormData, constants } from "js-forms";
 import { Button, Stack, Box, SimpleGrid } from "@chakra-ui/core";
-import valSchema from "./register-form-validation";
-import CommonFormikInput from "./CommonFormikInput";
-import { hasErrors } from "./ModelValidationSummary";
-import PasswordRequirementsInfo from "./PasswordRequirementsInfo";
+import valSchema from "../register-form-validation";
+import { hasErrors } from "../../../../components/ModelValidationSummary";
+import CommonFormikInput from "../../../../components/CommonFormikInput";
+import PasswordRequirementsInfo from "../../../../components/PasswordRequirementsInfo";
 
 const RegisterForm = ({ ModelState, FullName, Email, EmailConfirm }) => {
   const [hideEmailConfirm, setHideEmailConfirm] = useState(
