+++ src/Directory/Directory/Pages/js/pages/Account/Register/index.js
@@ -1,0 +1,2 @@
+export { default as Register } from "./Register";
+export { default as RegisterResult } from "./RegisterResult";
