+++ src/Directory/Directory/Pages/js/pages/Account/index.js
@@ -1,0 +1,4 @@
+export * from "./Confirm";
+export * from "./Login";
+export * from "./Logout";
+export * from "./Register";
