+++ src/Directory/Directory/Pages/js/routes.js
@@ -1,12 +1,14 @@
 ﻿import React from "react";
-import Login from "../Account/pages/Login";
-import LoginRedirect from "../Account/pages/LoginRedirect";
-import ConfirmLogout from "../Account/pages/ConfirmLogout";
-import LogoutRedirect from "../Account/pages/LogoutRedirect";
-import Register from "../Account/pages/Register";
-import RegisterResult from "../Account/pages/RegisterResult";
-import ConfirmAccount from "../Account/pages/ConfirmAccount";
-import ResendConfirm from "../Account/pages/ResendConfirm";
+import {
+  Login,
+  LoginRedirect,
+  ConfirmLogout,
+  LogoutRedirect,
+  Register,
+  RegisterResult,
+  ConfirmAccount,
+  ResendConfirm
+} from "./pages/Account";
 
 // Here we map "page" components to keys
 export default {
